Create celery task for re-aggregating a plan. Closes #215
Also added 'is_reaggregating' column to Plan model, so
it is apparent when a plan is in the process of re-
aggregation. This is needed for plan chooser display.

When a plan is re-aggregating, 'is_reaggregating' is
set to True. When it is finished, it is set to False,
and 'is_pending' is also set to False. 'is_pending' is
set on a plan when it needs re-aggregation. This needs
to occur whenever subjects are modified.