Add jQuery-compatible object for encapsulating reaggregation
Also, colorize grid rows based on reaggregation state. The
reaggregator module doesn't yet contain the logic for allowing
the user to begin a reaggregation. What it does have currently
is a timer for checking the reaggregation state of displayed
plan rows, and the logic for acting on state changes.

Things still left to do include adding a reaggregate button to
the interface when the user is allowed to reaggregate the plan,
and adjusting any appropriate text. Also needed is a REST
endpoint for reaggregation.