Break config out into its own package and do away with config container
This was necessary to get languages working right and was something we
were going to do anyway eventually. Having configuration in a
pip-installable package also makes things a simpler in the sense that
the django container already comes pre-configured as a step in its
Dockerfile. It's also nice to separate out the config dependencies.

Currently the package lives inside `django/publicmapping/config` but it
will be easy to separate it out.