Exclude xmlconfig.po from Vagrant rsync
`xmlconfig.po` is derived from `config.xml`. As such, it should not be
put into the VM where it overrides the `xmlconfig.po` file generated by
the `django` container.

There was a bug that manifested itself as follows: whenever a file was
changed with `vagrant rsync-auto` running, `xmlconfig.po` within the VM
would be overwritten. If you would re-configure by running `./manage.py
setup config/config.xml`, the file would then be generated correctly,
but any other change to a file with rysnc-auto running would overwrite
it again.

This change makes it so that other `.po` files that are committed to
source control with proper translations (such as `django.po`) will still
make their way into the VM while `xmlconfig.po` will not.

A change to fix globbing to ignore `.mo` files in nested subfolders is
also included.