Remove specific references to PA instance from infrastructure config (#131)
* Remove references to DistrictBuilder PA from infrastructure

In bringing over some infrastructure updates from the PA instance
of DistrictBuilder, we introduced a few references to variables in
the PA instance infrastructure. Either make these generic, or revert
them to refer to Virginia where appropriate.

* Remove CloudFormation template

Now that the deployment infrastructure is fully migrated to
Terraform, there's no reason to keep a CloudFormation template
file in this repository.